Rewrite of core code to remove GMP dependency
This is a complete rewrite of the initial code for a few purposes. The
first is to remove the dependency on GMP and instead used a custom
256-bit memory block.

The second reason is to revise some logic problems and correct them.
Alongside this is the removal of global variables and them being placed
in a custom gamestate.