Alter timing to not be fixed to frame timings
A problem in other tetris games is that final times are calculated from
the frame count, so we only get an accuracy of 1/60th of a minute, which
means that we do encounter repeated timings every so often.

Changing to use the nanosecond counter means two identical times will
likely never occur.