Add X11 graphical version.
This version uses much less memory than other graphical versions
and could be considered the best default for any system which utilize an
x server.

This does not perform any caching or drawing, so it is quite unoptimized
in that respect. Caching of the field state in the graphical context may
be done, since it is simple and would have some decent performance
improvements.