Add some more warning flags to test compilation
These warning flags are useful in finding various errors. Probably most
important is the usage of the fsanitize flag, which for this commit
indicates the presence of udefined-behaviour in the form of a 64-bit
shift.

This error will be fixed in the next commit, so this build will still
fail.